pyproject.toml
[project]
name = "focus-time"
version = "0.1.0"
description = "focus time exp"
authors = [
    {name = "Nemanja Cukaric", email = "nemanja@logikka.ai"},
]

dependencies = [
    "pandas>=2.2.3",
    "transformers>=4.46.2",
    "matplotlib>=3.9.2",
    "seaborn>=0.13.2",
    "ipykernel>=6.29.5",
    "openai>=1.54.4",
    "PyYAML>=6.0.2",
    "tqdm>=4.67.0",
    "scikit-learn>=1.5.2",
    "torch==2.5.1+cu118",
    "numpy>=2.1.3",
    "plotly>=5.24.1",
    "nbformat>=5.10.4",
    "ipywidgets>=8.1.5",
    "bertviz>=1.4.0",
    "loguru>=0.7.2",
    "tensorboard>=2.18.0",
    "torch-tb-profiler>=0.4.3",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[[tool.pdm.source]]
url = "https://download.pytorch.org/whl/cu118"
name = "pytorch"
type = "index"

[[tool.pdm.source]]
url = "https://pypi.org/simple"
name = "pypi"
type = "index"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true
resolution.backend = "uv"

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'

[tool.pdm.dev-dependencies]
dev = [
    "pre-commit>=3.5.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.10.0",
    "pre-commit>=4.0.1",
]


